# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Deploy containers to cloud servers. Push a config YAML + env, get an ordered deployment plan, execute it.",
        "title": "nvoi API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/workspaces/{workspace_id}/repos/{repo_id}/describe": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns live cluster state (nodes, workloads, pods, services, ingress, secrets, storage) via SSH. Same data as `nvoi describe` in direct mode.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Describe cluster",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Repo ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_core.DescribeResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/workspaces/{workspace_id}/repos/{repo_id}/resources": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all resources across configured providers (compute, DNS, storage). Same data as `nvoi resources` in direct mode.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "List provider resources",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Repo ID",
                        "name": "repo_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_provider.ResourceGroup"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_getnvoi_nvoi_pkg_core.DescribeIngress": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "github_com_getnvoi_nvoi_pkg_core.DescribeNode": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_getnvoi_nvoi_pkg_core.DescribePod": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "node": {
                    "type": "string"
                },
                "restarts": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_getnvoi_nvoi_pkg_core.DescribeResult": {
            "type": "object",
            "properties": {
                "ingress": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_core.DescribeIngress"
                    }
                },
                "namespace": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_core.DescribeNode"
                    }
                },
                "pods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_core.DescribePod"
                    }
                },
                "secrets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_core.DescribeSecret"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_core.DescribeService"
                    }
                },
                "storage": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_core.StorageItem"
                    }
                },
                "workloads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_getnvoi_nvoi_pkg_core.DescribeWorkload"
                    }
                }
            }
        },
        "github_com_getnvoi_nvoi_pkg_core.DescribeSecret": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "description": "obfuscated",
                    "type": "string"
                }
            }
        },
        "github_com_getnvoi_nvoi_pkg_core.DescribeService": {
            "type": "object",
            "properties": {
                "cluster_ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ports": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "github_com_getnvoi_nvoi_pkg_core.DescribeWorkload": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "kind": {
                    "description": "\"deployment\" or \"statefulset\"",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ready": {
                    "description": "\"2/2\"",
                    "type": "string"
                }
            }
        },
        "github_com_getnvoi_nvoi_pkg_core.StorageItem": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_getnvoi_nvoi_pkg_provider.ResourceGroup": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "internal_api_handlers.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "workspace not found"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token from POST /login. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}